requirements.txt
requests==2.32.3
python-dotenv==1.1.0
rich == 14.0.0
